+++ src/signals/incident/components/form/ContainerSelect/LegendPanel/LegendPanel.test.tsx
