+++ src/signals/incident/components/form/ContainerSelect/LegendPanel/LegendPanel.tsx
