+++ src/signals/incident/components/form/ContainerSelect/LegendPanel/index.ts
@@ -1,0 +1,3 @@
+import LegendPanel from './LegendPanel';
+
+export default LegendPanel;
