+++ src/signals/incident/components/form/ContainerSelect/ViewerContainer/index.ts
@@ -1,0 +1,3 @@
+import ViewerContainer from './ViewerContainer';
+
+export default ViewerContainer;
